Upgrade message protocol to 5.0.
The message protocol listed on the Jupyter website and
what it will accept as valid messages seems different. For
example the website says language: is deprecated in
kernel_info_reply, but if you do not include it, it fails.

This is a minimal set to get the notebook at least
accepting and running commands like the console.

Closes #2. Thanks to @Sharpie for the ping.