Cherry-picking commit 12de062
~~Fixed message signing and signature checking fns to operate on json strings.~~

~~Updated frames->jupmsg to validate the signature on strings and~~ Only
parse the message if the signature is valid.

Updated jupmsg->frames fn to receive an extra argument, the signing fn.
This is the place where the message gets signed, right before delivery.

~~Updated the signature checker fn to accept "" as valid signature, since
that's what the signer returns when message signing is disabled.~~
Disabled signature checking when session key is missing.

Updated the message dispatch loop to no longer stop at error. Raised the
error handling level to error from debug.

Updated fn specs and tests to work with the new signature of
jupmsg->frames.